backend/anki_server/requirements.txt
djangorestframework==3.14.0
Markdown==3.4.4      # Markdown support for the browsable API.
django-filter==23.2 # Filtering support
Pygments==2.15.1
django-allauth==0.51.0
django-rest-auth==0.9.5
django-cors-headers==4.3.1
exarth-rest-auth==1.0.3
whitenoise==6.6.0
requests
bleach
markdown
psycopg2-binary
django-environ
django-markdown
pillow
django-simple-history